refactor: split agents controller into crud + observability (#2253)
